+++ IndexTextCell.m
@@ -56,7 +56,7 @@
 /*
 - (void)drawInteriorWithFrame:(NSRect)cellFrame inView:(NSView *)controlView
 {
-	int textHeight;
+	CGFloat textHeight;
 	
 	NSRect inset = cellFrame;
 	NSMutableDictionary *attrs;
